Include undo functionality
There is no way for the user to revert a command if they have made a
mistake.

To allow the user to revert to a previous stage, an "undo" feature was
added, by putting the different states into an array list and adjusting
the pointer accordingly if an undo command is called.

This allows the different states of the tasklist to be saved easily.